apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0
celery==5.3.4
